littlehelp/phase3/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
beautifulsoup4==4.12.3
playwright==1.52.0
anthropic>=0.39.0
python-dotenv==1.0.1
pydantic==2.9.0
